bare rearrangement 2
